Extract ThreadHeader + PENDING_ACTION_COPY from thread page
- Move the fixed thread header component to components/ThreadHeader.tsx
  so it's reusable from /thread/[id]/info and /thread/[id]/edit-title
  routes (they currently duplicate the same fixed-header pattern).
- Move PENDING_ACTION_COPY + PendingActionKind type into
  app/thread/[threadId]/threadActionCopy.ts. The constant table is
  module-local data, not page logic.

app/thread/[threadId]/page.tsx: 1972 -> 1875 lines.
